+++ api/types/plugin.go
@@ -1,5 +1,3 @@
-// +build experimental
-
 package types
 
 import (
