+++ daemon/graphdriver/plugin_unsupported.go
@@ -2,6 +2,8 @@
 
 package graphdriver
 
-func lookupPlugin(name, home string, opts []string) (Driver, error) {
+import "github.com/docker/docker/plugin/getter"
+
+func lookupPlugin(name, home string, opts []string, plugingetter getter.PluginGetter) (Driver, error) {
 	return nil, ErrNotSupported
 }
